requirements.txt
chainlit==2.9.5
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.4

